made the pivot Modal draggable (experimental), bug fixes with dates
Date conversion: The intention of the original fields, which show the first original value and a transformed value, was to provide an example of an original date and what it would look like. In situations where the date column was not clean, non-transformable values could be displayed. Because dates can still be converted, in this situation, the behaviour has been changed: Convertible original values are displayed as soon as the correct conversion string is entered.